Major milestone accomplished cURL is required for the SMTP support. Only gmail's SMTP server is currently supported. Launcher support for the SMTP-enabled release is not complete.
Currently, only sending a string as a message is supported.
Sending a file as a message gives a segfault on Windows (but not on Fedora 15)
That is surely a bug with phobos.

This module uses etc.curl which is not yet incorporated into Phobos completely.
Many thanks to jcd for his work concerning cURL and D.